requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
